requirements.txt
cudnn
jupyter
keras-applications
pillow
pyyaml
appnope
astor
chardet
cycler
imageio
kiwisolver
matplotlib
more-itertools
music21
networkx
opt-einsum
pandas
pydot
pydotplus
pytz
scikit-image
tensorflow-addons
testpath
typeguard
